Create OP_UPDATE_CLAIM, make OP_SUPPORT_CLAIM refer to claimID
Updates to claims now refer to a unique ID assigned to OP_CLAIM_NAMEs.
The claimID is the Hash160 of the claim's hash and nOut. When an
update is updated, that same claimID is used, so all updates to a
claim will use the original claimID.

Supports now refer to the claimID rather than a specific hash and
nOut.